% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_sample.R
\docType{data}
\name{vcf_sample}
\alias{vcf_sample}
\title{A VCF from this GWAS of Type 2 Diabetes https://doi.org/10.1038/s41588-018-0084-1. containing
a subset of 10,000 variants}
\format{
`vcf_sample`
A CollapsedVCF

dim: 10000 1
rowRanges(vcf):
  GRanges with 5 metadata columns: paramRangeID, REF, ALT, QUAL, FILTER
info(vcf):
  DataFrame with 1 column: AF
info(header(vcf)):
  Number Type  Description
AF A      Float Allele Frequency
geno(vcf):
  List of length 9: ES, SE, LP, AF, SS, EZ, SI, NC, ID
geno(header(vcf)):
  Number Type    Description
ES A      Float   Effect size estimate relative to the alternative allele
SE A      Float   Standard error of effect size estimate
LP A      Float   -log10 p-value for effect estimate
AF A      Float   Alternate allele frequency in the association study
SS A      Integer Sample size used to estimate genetic effect
EZ A      Float   Z-score provided if it was used to derive the EFFECT and SE fields
SI A      Float   Accuracy score of summary data imputation
NC A      Integer Number of cases used to estimate genetic effect
ID 1      String  Study variant identifier
}
\usage{
data("vcf_sample")
}
\description{
A VCF from this GWAS of Type 2 Diabetes https://doi.org/10.1038/s41588-018-0084-1. containing
a subset of 10,000 variants
}
\keyword{datasets}
