% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseControl_SE.R
\name{CaseControl_SE}
\alias{CaseControl_SE}
\title{CaseControl_SE}
\usage{
CaseControl_SE(N_case, N_control, OR, SE)
}
\arguments{
\item{N_case}{an integer of the number of Case individuals}

\item{N_control}{an integer of the number of Control individuals}

\item{OR}{a vector of odds ratios}

\item{SE}{a vector of standard errors for the OR calculation *make sure the indices match between the two vectors}
}
\value{
a dataframe with 3 columns: MAF_case, MAF_control, MAF_pop for the estimated MAFs for each variant
}
\description{
This is a function to derive the case and control AFs from GWAS summary statistics when
the user has access to the whole sample AF, the sample sizes, and the OR (or beta).
If user has SE instead of sample AF use [CaseControlAF::CaseControl_SE()]
This code uses the GroupFreq function adapted from C from <https://github.com/Paschou-Lab/ReAct/blob/main/GrpPRS_src/CountConstruct.c>
}
\note{
Make sure the vectors are in the same order for your variants
}
\examples{
data("sampleDat")

nCase_sample = 79148
nControl_sample = 61106

# get the estimated case and control MAFs
se_method_results <- CaseControl_SE(N_case = nCase_sample,
                                    N_control = nControl_sample,
                                    OR = sampleDat$OR,
                                    SE = sampleDat$SE)

head(se_method_results)

}
\references{
https://github.com/wolffha/CaseControlAF
}
\seealso{
\url{https://github.com/wolffha/CaseControlAF} for further documentation
}
\author{
Hayley Wolff (Stoneman), \email{hayley.wolff@cuanschutz.edu}
}
